// blacksmith-of-lenai-app/src/app/api-generated/model/jobType.ts
/**
 * Garden Game API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Die verfügbaren Berufe im Spiel
 */
export const JobType = {
    HerbCollector: 'HerbCollector',
    Lumberjack: 'Lumberjack',
    BlackSmith: 'BlackSmith',
    Miner: 'Miner'
} as const;
export type JobType = typeof JobType[keyof typeof JobType];